Add support for Chmod on billy.Filesystem
The billy package contains a `Change` interface type, which seems to
have gone unused for several years now, presumably due to the difficulty
of implementing most of the required methods for all supported
platforms.

This commit reduces the `Change` interface to a `Chmod` interface, which
is supported on all platforms. The interface is implemented in all
applicable abstractions.

Supporting `chmod` in billy would help with issues such as
https://github.com/go-git/go-git/issues/588 (which was closed as stale,
but is unfortunately still a real issue for some folks).

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>